⭐ Login Logout complete
~ the session persist around of the user cookie sesion
